+++ addons/es/tbai/invoice.go
@@ -20,6 +20,7 @@ var invoiceCorrectionDefinitions = tax.CorrectionSet{
 
 func validateInvoice(inv *bill.Invoice) error {
 	return validation.ValidateStruct(inv,
+    validation.Field(&inv.Series, validation.Required),
 		validation.Field(&inv.Customer,
 			validation.By(validateInvoiceCustomer),
 			validation.Skip,
@@ -37,6 +38,10 @@ func validateInvoice(inv *bill.Invoice) error {
 				validation.By(validateInvoiceLine),
 				validation.Skip,
 			),
+			validation.Skip,
+		),
+		validation.Field(&inv.Notes,
+			cbc.ValidateNotesHasKey(cbc.NoteKeyGeneral),
 			validation.Skip,
 		),
 	)
@@ -69,6 +74,7 @@ func validateInvoicePreceding(val any) error {
 	}
 	return validation.ValidateStruct(p,
 		validation.Field(&p.IssueDate, validation.Required),
+		validation.Field(&p.Series, validation.Required),
 		validation.Field(&p.Ext,
 			tax.ExtensionsRequires(ExtKeyCorrection),
 			validation.Skip,