+++ addons/es/tbai/invoices_test.go
@@ -5,11 +5,11 @@ import (
 
 	"github.com/invopop/gobl/addons/es/tbai"
 	"github.com/invopop/gobl/bill"
+	"github.com/invopop/gobl/cbc"
 	"github.com/invopop/gobl/num"
 	"github.com/invopop/gobl/org"
 	"github.com/invopop/gobl/tax"
 
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
@@ -42,17 +42,42 @@ func TestBasqueLineValidation(t *testing.T) {
 	assertValidationError(t, inv, "es-tbai-exemption: required")
 }
 
+func TestInvoiceValidation(t *testing.T) {
+	inv := validTicketBAIInvoice()
+	require.NoError(t, inv.Calculate())
+	require.NoError(t, inv.Validate())
+
+	inv.Lines[0].Taxes[0].Ext[tbai.ExtKeyProduct] = "services"
+	require.NoError(t, inv.Calculate())
+	require.NoError(t, inv.Validate())
+
+	inv.Lines[0].Taxes[0].Ext = nil
+	assertValidationError(t, inv, "es-tbai-exemption: required")
+
+	t.Run("without series", func(t *testing.T) {
+		inv := validTicketBAIInvoice()
+		inv.Series = ""
+		assertValidationError(t, inv, "series: cannot be blank")
+	})
+
+	t.Run("without notes", func(t *testing.T) {
+		inv := validTicketBAIInvoice()
+		inv.Notes = nil
+		assertValidationError(t, inv, "notes: with key 'general' missing")
+	})
+}
+
 func assertValidationError(t *testing.T, inv *bill.Invoice, expected string) {
 	require.NoError(t, inv.Calculate())
 	err := inv.Validate()
-	require.Error(t, err)
-	assert.Contains(t, err.Error(), expected)
+	require.ErrorContains(t, err, expected)
 }
 
 func testInvoiceStandard(t *testing.T) *bill.Invoice {
 	t.Helper()
 	return &bill.Invoice{
 		Addons: tax.WithAddons(tbai.V1),
+		Series: "ABC",
 		Code:   "123",
 		Supplier: &org.Party{
 			Name: "Test Supplier",
@@ -85,6 +110,12 @@ func testInvoiceStandard(t *testing.T) *bill.Invoice {
 						},
 					},
 				},
+			},
+		},
+		Notes: []*cbc.Note{
+			{
+				Key:  cbc.NoteKeyGeneral,
+				Text: "This is a test invoice",
 			},
 		},
 	}